+++ src/mapper.ts
@@ -1,43 +1,10 @@
-import type {
-  Atom,
-  Feed,
-  FeedEntry,
-  JsonFeed,
-  JsonFeedAuthor,
-  JsonFeedItem,
-  RSS1,
-  RSS2,
-} from "./types/mod.ts";
+import type { Feed, FeedEntry, JsonFeed } from "./types/mod.ts";
 import { isValidHttpURL } from "./util.ts";
 import { FeedType } from "./types/mod.ts";
 import { DublinCoreFieldArray, DublinCoreFields } from "./types/dublin-core.ts";
 import { InternalAtom } from "./types/internal-atom.ts";
 import { InternalRSS2 } from "./types/internal-rss2.ts";
 import { InternalRSS1 } from "./types/internal-rss1.ts";
-
-export const toJsonFeed = (
-  feedType: FeedType,
-  feed: Atom | RSS2 | RSS1,
-): JsonFeed | null => {
-  if (!feed) {
-    return null;
-  }
-
-  let result: JsonFeed | null = null;
-  switch (feedType) {
-    case FeedType.Atom:
-      result = mapAtomToJsonFeed(feed as Atom);
-      break;
-    case FeedType.Rss2:
-      result = mapRss2ToJsonFeed(feed as RSS2);
-      break;
-    case FeedType.Rss1:
-      const rss1 = feed as RSS1;
-      break;
-  }
-
-  return result;
-};
 
 export const toFeed = (
   feedType: FeedType,
@@ -59,150 +26,6 @@ export const toFeed = (
     default:
       return null;
   }
-};
-
-const mapRss2ToJsonFeed = (rss: RSS2): JsonFeed => {
-  const items = rss.channel?.items?.map((rssItem) => {
-    let authors, author, attachments;
-
-    if (rssItem.author) {
-      author = {
-        name: rssItem.author,
-      };
-      authors = [author];
-    } else if (rssItem[DublinCoreFields.Creator]) {
-      author = {
-        name: rssItem[DublinCoreFields.Creator]?.[0],
-      };
-      authors = rssItem[DublinCoreFields.Creator]?.map((creator) => ({
-        name: creator,
-      }));
-    }
-
-    if (rssItem.enclosure) {
-      attachments = [{
-        url: rssItem.enclosure.url,
-        mime_type: rssItem.enclosure.type,
-        size_in_bytes: rssItem.enclosure.length,
-      }];
-    }
-
-    return {
-      id: rssItem.guid,
-      summary: rssItem.description,
-      title: rssItem.title,
-      external_url: rssItem.link,
-      content_html: rssItem.link,
-      author,
-      authors,
-      attachments,
-      url: rssItem.link,
-      date_published: rssItem.pubDate,
-      date_publishedRaw: rssItem.pubDateRaw,
-      date_modified: rssItem.pubDate,
-      date_modifiedRaw: rssItem.pubDateRaw,
-    } as JsonFeedItem;
-  }) || [];
-
-  const channel = rss.channel;
-  let author, hubs;
-
-  if (channel && (channel.managingEditor || channel.webMaster)) {
-    author = {
-      url: channel.managingEditor || channel.webMaster,
-    };
-  }
-
-  if (channel?.cloud) {
-    hubs = [{
-      type: channel.cloud.protocol,
-      url: `${channel.cloud.domain}${
-        channel.cloud.port ? ":" + channel.cloud.port : ""
-      }${channel.cloud.path}`,
-    }];
-  }
-
-  return {
-    title: channel?.title,
-    description: channel?.description,
-    icon: channel?.image?.url,
-    home_page_url: channel?.link,
-    items,
-    author,
-    hubs,
-  } as JsonFeed;
-};
-
-const mapAtomToJsonFeed = (atom: Atom): JsonFeed => {
-  const feed = {
-    icon: atom.icon,
-    title: atom.title?.value ?? atom.title,
-    author: atom.author ? {
-			name: atom.author.name,
-      url: atom.author.uri,
-    } as JsonFeedAuthor : undefined,
-		items: atom.entries.map((entry) => {
-
-			let url: string | undefined;
-			if (entry["feedburner:origlink"]) {
-				url = entry["feedburner:origlink"];
-			} else if (entry.href) {
-				url = entry.href;
-			} else if (isValidHttpURL(entry.id)) {
-				url = entry.id;
-			}
-
-			const item: JsonFeedItem = {
-				id: entry.id,
-				title: entry.title?.value ?? entry.title,
-				date_published: entry.published,
-				date_publishedRaw: entry.publishedRaw,
-				date_modified: entry.updated,
-				date_modifiedRaw: entry.updatedRaw,
-				summary: entry.summary?.value,
-				tags: entry.categories?.map(x => x.term),
-				author: {
-					name: entry.author?.name,
-					url: entry.author?.uri,
-				},
-				url,
-				attachments: entry.links?.filter((link) => link.rel === "enclosure").map((link) => ({
-					url: link.href,
-					mime_type: link.type,
-					size_in_bytes: link.length,
-				})),
-			};
-
-			if (entry.content) {
-				switch (entry.content.type?.toUpperCase()) {
-					case "XHTML":
-					case "HTML":
-						item.content_html = entry.content.value;
-						break;
-					default:
-						item.content_text = entry.content.value;
-						break;
-				}
-			}
-
-			return item;
-		})
-  } as JsonFeed;
-
-  if (atom.links?.length) {
-    for (const link of atom.links) {
-      switch (link.rel) {
-        case "self":
-          feed.home_page_url = link.href;
-          break;
-        case "alternate":
-          feed.feed_url = link.href;
-          break;
-      }
-    }
-  }
-
-  return feed;
 };
 
 const mapRssToFeed = (rss: InternalRSS1): Feed => {
@@ -272,7 +95,7 @@ const mapRss2ToFeed = (rss: InternalRSS2): Feed => {
       result.createdRaw = createdRaw;
     }
 
-    result.links = [(rss.channel[DublinCoreFields.URI]?.value || (rss.channel.link.value as string))];
+    result.links = [(rss.channel[DublinCoreFields.URI]?.value || (rss.channel.link?.value as string))];
     result.language = rss.channel[DublinCoreFields.Language]?.value || rss.channel.language?.value;
     result.updateDate = rss.channel.lastBuildDate?.value;
     result.updateDateRaw = rss.channel.lastBuildDateRaw?.value;
@@ -420,8 +243,8 @@ const mapAtomToFeed = (atom: InternalAtom): Feed => {
     icon: atom.icon?.value,
     image: atom.logo
       ? {
-        link: atom.logo?.value,
-        url: atom.logo?.value,
+        link: atom.logo.value,
+        url: atom.logo.value,
       }
       : undefined,
     updateDate: atom.updated?.value,
